move App.js code to match each components
